cleanup/refactor __init__
